mini_pjt/mini-pjt09/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.2.9
idna==3.2
python-dotenv==0.19.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.6
